[maint] Rename activity status to step
