+++ mspelling/lib/activity/spelling_controller.dart
@@ -2,7 +2,7 @@ import 'package:get/get.dart';
 import 'package:data/db.dart';
 import 'package:mspelling/login/login_controller.dart';
 import 'package:mspelling/setup/setup_controller.dart';
-import 'package:mspelling/activity/status.dart';
+import 'package:mspelling/activity/step.dart';
 import 'package:mspelling/activity/stim_controller.dart';
 import 'package:mspelling/errors/errors.dart';
